backend/requirements.txt
fastapi~=0.123.10
uvicorn~=0.38.0
requests~=2.32.5
aiohttp

SQLAlchemy~=2.0.44
alembic~=1.17.2

bcrypt~=4.0.1
pydantic~=2.12.5
python-dotenv~=1.2.1
passlib~=1.7.4
python-jose~=3.5.0